add g function support, merging cfour tests into main file
