--- iodata/test/data/h_donly_sph.molden
@@ -1,110 +1,0 @@
-[Molden Format]
-[ATOMS] AU
-H        1    1        0.0000000000        0.0000000000        0.0000000000
-[GTO]
-  1 0
-d   1 1.00
-  0.500000000000000        1.00000000000000
-
-
-[MO]
- Sym= A
- Ene=   1.14819777754906
- Spin= Alpha
- Occup= 0.0
-  1       -0.2886751346
-  2       -0.2886751346
-  3        0.5773502692
-  4        0.0000000000
-  5        0.0000000000
-  6        0.0000000000
- Sym= A
- Ene=   1.14819777754906
- Spin= Alpha
- Occup= 0.0
-  1        0.5000000000
-  2       -0.5000000000
-  3        0.0000000000
-  4        0.0000000000
-  5        0.0000000000
-  6        0.0000000000
- Sym= A
- Ene=   1.14819777754906
- Spin= Alpha
- Occup= 0.0
-  1        0.0000000000
-  2        0.0000000000
-  3        0.0000000000
-  4        1.0000000000
-  5        0.0000000000
-  6        0.0000000000
- Sym= A
- Ene=   1.14819777754906
- Spin= Alpha
- Occup= 0.0
-  1        0.0000000000
-  2        0.0000000000
-  3        0.0000000000
-  4        0.0000000000
-  5        1.0000000000
-  6        0.0000000000
- Sym= A
- Ene=   1.14819777754906
- Spin= Alpha
- Occup= 0.0
-  1        0.0000000000
-  2        0.0000000000
-  3        0.0000000000
-  4        0.0000000000
-  5        0.0000000000
-  6        1.0000000000
- Sym= A
- Ene=   1.14819777754906
- Spin= Beta
- Occup= 0.0
-  1       -0.2886751346
-  2       -0.2886751346
-  3        0.5773502692
-  4        0.0000000000
-  5        0.0000000000
-  6        0.0000000000
- Sym= A
- Ene=   1.14819777754906
- Spin= Beta
- Occup= 0.0
-  1        0.5000000000
-  2       -0.5000000000
-  3        0.0000000000
-  4        0.0000000000
-  5        0.0000000000
-  6        0.0000000000
- Sym= A
- Ene=   1.14819777754906
- Spin= Beta
- Occup= 0.0
-  1        0.0000000000
-  2        0.0000000000
-  3        0.0000000000
-  4        1.0000000000
-  5        0.0000000000
-  6        0.0000000000
- Sym= A
- Ene=   1.14819777754906
- Spin= Beta
- Occup= 0.0
-  1        0.0000000000
-  2        0.0000000000
-  3        0.0000000000
-  4        0.0000000000
-  5        1.0000000000
-  6        0.0000000000
- Sym= A
- Ene=   1.14819777754906
- Spin= Beta
- Occup= 0.0
-  1        0.0000000000
-  2        0.0000000000
-  3        0.0000000000
-  4        0.0000000000
-  5        0.0000000000
-  6        1.0000000000
