+++ iodata/test/data/h_sonly_sph_cfour.molden
