+++ iodata/test/test_molden.py
@@ -207,6 +207,29 @@ def test_load_molden_nh3_molden_cart():
     molden_charges = np.array([0.3138, -0.4300, -0.0667, 0.1829])
     assert_allclose(charges, molden_charges, atol=1.e-3)
 
+def test_load_molden_h_cfour():
+    # The file tested here is created with CFOUR 2.1.
+    file_list = [
+                 'h_sonly_sph.molden',
+                 'h_ponly_sph.molden',
+                 'h_donly_sph.molden',
+                 'h_fonly_sph.molden',
+                 'h_gonly_sph.molden',
+                 'h_sonly_cart.molden',
+                 'h_ponly_cart.molden',
+                 'h_donly_cart.molden',
+                 'h_fonly_cart.molden',
+                 'h_gonly_cart.molden',
+                ]
+
+    for i in file_list:
+        with path('iodata.test.data', i) as fn_molden:
+            print(str(fn_molden))
+            mol = load_one(str(fn_molden))
+            # Check normalization
+            olp = compute_overlap(mol.obasis, mol.atcoords)
+            check_orthonormal(mol.mo.coeffsa, olp)
+            check_orthonormal(mol.mo.coeffsb, olp)
 
 def test_load_molden_nh3_orca():
     # The file tested here is created with ORCA. It should be read in
